Add PubSub test setup file with env adapter testing
